backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# DICOM processing
pydicom==3.0.2
pyjpegls>=1.2          # required for JPEG-LS Lossless DICOM decompression (used by CRE)
numpy>=1.26.4
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# ML / PyTorch  (install CUDA version separately: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118)
torch>=2.2.0
torchvision>=0.17.0
scikit-learn==1.4.2
scipy==1.13.0
matplotlib==3.8.4

# RAG / Embeddings
chromadb==0.5.0
sentence-transformers==2.7.0

# Ollama client
ollama==0.2.1

# PDF generation
reportlab==4.1.0
PyPDF2==3.0.1

# PDF text extraction (for parsing training compte rendus)
pdfplumber==0.11.0

# MedSAM pseudo-labeling (optional — only needed for annotation pipeline)
# Install with: pip install transformers accelerate
# OR:           pip install segment-anything  (+ download sam_vit_b_01ec64.pth manually)
transformers>=4.40.0
accelerate>=0.29.0

# Utils
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
